refactor: keep chart renderer within quality budget
