+++ package.json
@@ -26,7 +26,7 @@
     "test:visualization-ir": "bun test web/components/dashboard/visualization/ir-contract.test.ts web/components/dashboard/visualization/host.test.ts web/components/dashboard/visualization/format.test.ts web/components/dashboard/visualization/conditional-format.test.ts web/components/dashboard/visualization/metadata.test.ts web/components/dashboard/visualization/interaction-command.test.ts web/components/dashboard/visualization/signal-envelope.test.ts web/components/dashboard/visualization/layout.test.ts web/components/dashboard/visualization/adapters/*.test.ts web/components/dashboard/visualization/adapters/maplibre/spatial-selection.test.ts web/components/dashboard/table/format.test.ts web/components/dashboard/table/conditional-formatting.test.ts",
     "test:app-shell": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts app-shell && bun test web/components/app/app-shell.dom.test.ts",
     "test:table-selection": "bun test web/components/dashboard/table/block-source.test.ts web/components/dashboard/table/selection.test.ts",
-    "test:interaction-selection": "bun test web/components/dashboard/interaction-selection.test.ts web/components/dashboard/visualization/interaction-command.test.ts web/components/dashboard/visualization/adapters/echarts.test.ts web/components/dashboard/visualization/adapters/echarts-point.test.ts web/components/dashboard/visualization/adapters/maplibre.test.ts web/components/dashboard/table/selection.test.ts web/components/dashboard/filters/filter-controller.test.ts",
+    "test:interaction-selection": "bun test web/components/dashboard/interaction-selection.test.ts web/components/dashboard/visualization/interaction-command.test.ts web/components/dashboard/visualization/adapters/echarts.test.ts web/components/dashboard/visualization/adapters/echarts-series-intent.test.ts web/components/dashboard/visualization/adapters/echarts-point.test.ts web/components/dashboard/visualization/adapters/maplibre.test.ts web/components/dashboard/table/selection.test.ts web/components/dashboard/filters/filter-controller.test.ts",
     "test:movielens-performance-harness": "bun test scripts/movielens_performance.test.ts",
     "test:public-site-smoke": "bun test scripts/public_site_smoke.test.ts",
     "test:visual-modal": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts visual-modal && bun test web/components/dashboard/visual-modal-focus.test.ts web/components/dashboard/visual-modal.dom.test.ts",
