+++ web/components/dashboard/visualization/adapters/echarts.test.ts
@@ -491,282 +491,6 @@ test('ECharts translation applies combo marks and axes to multi-measure series',
   expect(temporalHorizontalOption.xAxis[1]).toMatchObject({ type: 'time', inverse: true, splitNumber: 3 })
 })
 
-test('ECharts normalizes stacks and preserves series order and color identity across filters', () => {
-  const envelope = cartesianSeriesFixture() as any
-  envelope.spec.presentation.stacking = 'percent'
-  envelope.spec.presentation.seriesIntent = [
-    { value: 'delivered', order: 0, color: 'success' },
-    { value: 'processing', order: 1, color: 'data_3' },
-    { value: 'canceled', order: 2, color: 'danger' },
-  ]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  expect(option.series.map((series: any) => series.name)).toEqual(['delivered', 'processing'])
-  expect(option.series.map((series: any) => series.itemStyle.color)).toEqual([
-    defaultRendererContext.colors.success,
-    defaultRendererContext.colors.data[2],
-  ])
-  expect(option.series.map((series: any) => series.stack)).toEqual(['percent', 'percent'])
-  expect(option.series.map((series: any) => series.encode.y)).toEqual(['__lv_percent_value', '__lv_percent_value'])
-  expect(option.dataset[1].source).toEqual([
-    ['label', 'series', 'value', '__lv_percent_value'],
-    ['Jan', 'delivered', 10, 25],
-    ['Feb', 'delivered', 30, 75],
-  ])
-  expect(option.dataset[2].source).toEqual([
-    ['label', 'series', 'value', '__lv_percent_value'],
-    ['Jan', 'processing', 30, 75],
-    ['Feb', 'processing', 10, 25],
-  ])
-  expect(option.yAxis.axisLabel.formatter(25)).toBe('25%')
-
-  const darkContext = {
-    ...defaultRendererContext,
-    theme: 'dark',
-    colors: { ...defaultRendererContext.colors, success: '#2ea043', data: ['#1f6feb', '#a371f7', '#d29922'] },
-  } as any
-  const darkOption = echartsOption(envelope, darkContext) as any
-  expect(darkOption.series.map((series: any) => series.itemStyle.color)).toEqual(['#2ea043', '#d29922'])
-
-  const filtered = structuredClone(envelope) as any
-  filtered.dataState.datasets[0].rows = filtered.dataState.datasets[0].rows.filter((row: unknown[]) => row[1] === 'processing')
-  const filteredOption = echartsOption(filtered, defaultRendererContext) as any
-  expect(filteredOption.series.map((series: any) => [series.name, series.itemStyle.color])).toEqual([
-    ['processing', defaultRendererContext.colors.data[2]],
-  ])
-  expect(filteredOption.dataset[1].source.map((row: unknown[]) => row.at(-1))).toEqual(['__lv_percent_value', 100, 100])
-
-  delete envelope.spec.presentation.seriesIntent
-  const categoryColors = new CategoryColorRegistry()
-  const automatic = echartsOption(envelope, defaultRendererContext, categoryColors) as any
-  const automaticProcessing = automatic.series.find((series: any) => series.name === 'processing').itemStyle.color
-  delete filtered.spec.presentation.seriesIntent
-  const automaticFiltered = echartsOption(filtered, defaultRendererContext, categoryColors) as any
-  expect(automaticFiltered.series[0].itemStyle.color).toBe(automaticProcessing)
-})
-
-test('ECharts uses governed static colors for category series and their legends', () => {
-  const envelope = cartesianSeriesFixture() as any
-  envelope.spec.conditionalFormatting = [{
-    id: 'status-colors', target: 'series_color', field: { dataset: 'primary', field: 'value' },
-    rule: {
-      kind: 'field', source: { dataset: 'primary', field: 'series' },
-      values: {
-        delivered: { color: 'data_1', icon: 'circle' },
-        processing: { color: 'data_6', icon: 'circle' },
-      },
-      nullStyle: { color: 'neutral', icon: 'square' },
-      defaultStyle: { color: 'neutral', icon: 'square' },
-    },
-  }]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  expect(option.series.map((series: any) => [series.name, series.itemStyle.color])).toEqual([
-    ['delivered', defaultRendererContext.colors.data[0]],
-    ['processing', defaultRendererContext.colors.data[5]],
-  ])
-
-  envelope.dataState.datasets[0].rows = envelope.dataState.datasets[0].rows.filter((row: unknown[]) => row[1] === 'processing')
-  const filtered = echartsOption(envelope, defaultRendererContext) as any
-  expect(filtered.series[0].itemStyle.color).toBe(defaultRendererContext.colors.data[5])
-})
-
-test('ECharts keeps typed category-series identities distinct through ordering and filters', () => {
-  const envelope = cartesianSeriesFixture() as any
-  envelope.dataState.datasets[0].rows = [
-    ['Jan', 1, 10], ['Jan', '1', 20], ['Feb', 1, 30], ['Feb', '1', 40],
-  ]
-  envelope.spec.presentation.seriesIntent = [{ value: '1', order: 0, color: 'success' }]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  const series = option.series.filter((candidate: any) => candidate.datasetId)
-  expect(series.map((candidate: any) => candidate.name)).toEqual(['1 [string:1]', '1 [number:1]'])
-  expect(new Set(series.map((candidate: any) => candidate.id)).size).toBe(2)
-  expect(new Set(series.map((candidate: any) => candidate.datasetId)).size).toBe(2)
-  expect(series.map((candidate: any) => candidate.itemStyle.color)).toEqual([
-    defaultRendererContext.colors.success,
-    defaultRendererContext.colors.data[0],
-  ])
-  expect(option.dataset.slice(1).map((dataset: any) => dataset.transform.config['='])).toEqual(['1', 1])
-  const reordered = structuredClone(envelope)
-  reordered.dataState.datasets[0].rows.reverse()
-  expect((echartsOption(reordered, defaultRendererContext) as any).series.filter((candidate: any) => candidate.datasetId).map((candidate: any) => candidate.id)).toEqual(series.map((candidate: any) => candidate.id))
-
-  const canonical = structuredClone(envelope)
-  canonical.spec.presentation.seriesIntent = [{ value: '1 [number:1]', order: 0, color: 'danger' }]
-  const canonicalOption = echartsOption(canonical, defaultRendererContext) as any
-  expect(canonicalOption.series.filter((candidate: any) => candidate.datasetId).map((candidate: any) => candidate.name)).toEqual(['1 [number:1]', '1 [string:1]'])
-  expect(canonicalOption.dataset.slice(1).map((dataset: any) => dataset.transform.config['='])).toEqual([1, '1'])
-
-  envelope.spec.datasets[0].fields[1].role = 'identity'
-  envelope.spec.interactions = [{
-    id: 'point_selection', kind: 'select', mode: 'single', requiresStableIdentity: true, targets: ['details'], mappings: [
-      { source: { dataset: 'primary', field: 'series' }, targetFieldID: 'orders.series', targetDatasetID: 'orders' },
-    ],
-  }]
-  expect(interactionCommandForRow(envelope, 'primary', envelope.dataState.datasets[0].rows[0])).toMatchObject({ mappings: [{ value: 1 }] })
-  expect(interactionCommandForRow(envelope, 'primary', envelope.dataState.datasets[0].rows[1])).toMatchObject({ mappings: [{ value: '1' }] })
-})
-
-test('ECharts preserves nullish category-series identities and raw filter values', () => {
-  const envelope = cartesianSeriesFixture() as any
-  envelope.dataState.datasets[0].rows = [
-    ['Jan', null, 10], ['Jan', undefined, 20], ['Feb', null, 30], ['Feb', undefined, 40],
-  ]
-  // Nullish values are authored through the renderer's display names because
-  // seriesIntent values are strings in the visualization contract.
-  envelope.spec.presentation.seriesIntent = [
-    { value: '(undefined)', order: 0, color: 'success' },
-    { value: '(null)', order: 1, color: 'danger' },
-  ]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  const series = option.series.filter((candidate: any) => candidate.datasetId)
-  expect(series.map((candidate: any) => candidate.name)).toEqual(['(undefined)', '(null)'])
-  expect(new Set(series.map((candidate: any) => candidate.id)).size).toBe(2)
-  expect(new Set(series.map((candidate: any) => candidate.datasetId)).size).toBe(2)
-  expect(series.map((candidate: any) => candidate.itemStyle.color)).toEqual([
-    defaultRendererContext.colors.success,
-    defaultRendererContext.colors.danger,
-  ])
-  expect(option.dataset.slice(1).map((dataset: any) => dataset.transform.config['='])).toEqual([undefined, null])
-})
-
-test('ECharts lets a single numeric category resolve a string series intent', () => {
-  const envelope = cartesianSeriesFixture() as any
-  envelope.dataState.datasets[0].rows = [['Jan', 7, 10], ['Feb', 7, 20]]
-  envelope.spec.presentation.seriesIntent = [{ value: '7', order: 0, color: 'danger' }]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  const series = option.series.find((candidate: any) => candidate.datasetId)
-  expect(series.name).toBe('7')
-  expect(series.itemStyle.color).toBe(defaultRendererContext.colors.danger)
-  expect(option.dataset[1].transform.config['=']).toBe(7)
-})
-
-test('ECharts keeps fallback multi-metric colors bound to fields when order changes', () => {
-  const envelope = cartesianFixture('line', ['label', 'revenue', 'cost']) as any
-  const initial = echartsOption(envelope, defaultRendererContext) as any
-  const initialColors = new Map(initial.series.map((series: any) => [series.name, series.itemStyle.color]))
-
-  envelope.spec.presentation.seriesIntent = [
-    { value: 'cost', order: 0 },
-    { value: 'revenue', order: 1 },
-  ]
-  const reordered = echartsOption(envelope, defaultRendererContext) as any
-  expect(reordered.series.map((series: any) => series.name)).toEqual(['cost', 'revenue'])
-  expect(reordered.series.map((series: any) => [series.name, series.itemStyle.color])).toEqual([
-    ['cost', initialColors.get('cost')], ['revenue', initialColors.get('revenue')],
-  ])
-})
-
-test('ECharts places orderless static intents before unconfigured measures', () => {
-  const envelope = cartesianFixture('line', ['label', 'revenue', 'cost']) as any
-  envelope.spec.presentation.seriesIntent = [{ value: 'cost' }]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  expect(option.series.map((series: any) => series.name)).toEqual(['cost', 'revenue'])
-  expect(option.series.map((series: any) => series.itemStyle.color)).toEqual([
-    defaultRendererContext.colors.data[1], defaultRendererContext.colors.data[0],
-  ])
-})
-
-test('ECharts preserves conditional-over-intent-over-palette precedence for category series', () => {
-  const envelope = cartesianSeriesFixture() as any
-  envelope.spec.presentation.seriesIntent = [{ value: 'delivered', color: 'success' }, { value: 'processing', color: 'warning' }]
-  envelope.spec.conditionalFormatting = [{
-    id: 'status-colors', target: 'series_color', field: { dataset: 'primary', field: 'value' },
-    rule: {
-      kind: 'field', source: { dataset: 'primary', field: 'series' },
-      values: { delivered: { color: 'danger' } },
-      nullStyle: {}, defaultStyle: {},
-    },
-  }]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  const processing = option.series.find((series: any) => series.name === 'processing').itemStyle.color
-  expect(option.series.find((series: any) => series.name === 'delivered').itemStyle.color).toBe(defaultRendererContext.colors.danger)
-  expect(typeof processing === 'function' ? processing({ value: ['Jan', 'processing', 30] }) : processing).toBe(defaultRendererContext.colors.attention)
-})
-
-test('ECharts condenses crowded category-series charts without overlapping labels or legends', () => {
-  const envelope = cartesianSeriesFixture() as any
-  envelope.spec.mark = 'column'
-  envelope.spec.presentation.stacked = false
-  envelope.spec.presentation.stacking = 'none'
-  const states = ['MG', 'CE', 'GO', 'MT', 'PE', 'RJ', 'RS', 'AP']
-  const statuses = ['approved', 'canceled', 'created', 'delivered', 'invoiced', 'processing', 'shipped', 'unavailable']
-  envelope.dataState.datasets[0].rows = states.flatMap((state: string, stateIndex: number) =>
-    statuses.map((status: string, statusIndex: number) => [state, status, (stateIndex + 1) * (statusIndex + 1)]),
-  )
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  expect(option.legend).toMatchObject({ type: 'scroll', orient: 'horizontal', left: 8, right: 8, height: 24, bottom: 0 })
-  expect(option.grid).toMatchObject({ left: 12, right: 16, top: 16, bottom: 44, containLabel: true })
-  expect(option.yAxis).toMatchObject({ splitNumber: 4, axisLabel: { hideOverlap: true } })
-  expect(option.series).toHaveLength(8)
-  expect(option.series.every((series: any) => series.label.show === false)).toBe(true)
-
-  const compact = echartsOption(cartesianSeriesFixture(), defaultRendererContext) as any
-  expect(compact.legend.type).toBeUndefined()
-  expect(compact.series.every((series: any) => series.label.show === true)).toBe(true)
-})
-
-test('ECharts normalizes multi-metric percent stacks without changing raw tooltip values', () => {
-  const envelope = cartesianFixture('area', ['label', 'revenue', 'cost']) as any
-  envelope.spec.presentation.stacked = false
-  envelope.spec.presentation.stacking = 'percent'
-  envelope.spec.presentation.labelPolicy.density = 'hidden'
-  envelope.spec.presentation.seriesIntent = [
-    { value: 'cost', order: 0, color: 'warning' },
-    { value: 'revenue', order: 1, color: 'success' },
-  ]
-  envelope.dataState.datasets[0].rows = [
-    ['Jan', 10, 30],
-    ['Feb', 30, 10],
-  ]
-
-  const option = echartsOption(envelope, defaultRendererContext) as any
-  expect(option.series.map((series: any) => series.name)).toEqual(['cost', 'revenue'])
-  expect(option.series.map((series: any) => series.encode.y)).toEqual(['__lv_percent_cost', '__lv_percent_revenue'])
-  expect(option.series.map((series: any) => series.itemStyle.color)).toEqual([
-    defaultRendererContext.colors.attention,
-    defaultRendererContext.colors.success,
-  ])
-  expect(option.series.every((series: any) => series.label.show === false)).toBe(true)
-  expect(option.series[0].label.formatter({ value: ['Jan', 10, 30, 75, 25] })).toBe('75%')
-  expect(option.series[1].label.formatter({ value: ['Jan', 10, 30, 75, 25] })).toBe('25%')
-  expect(option.dataset.source).toEqual([
-    ['label', 'revenue', 'cost', '__lv_percent_cost', '__lv_percent_revenue'],
-    ['Jan', 10, 30, 75, 25],
-    ['Feb', 30, 10, 25, 75],
-  ])
-})
-
-test('ECharts translation emits one multi-value financial series', () => {
-  const envelope = {
-    schemaVersion: 9, visualID: 'ohlc', rendererID: 'echarts', specRevision: 'sha256:test', dataRevision: 1,
-    spec: {
-      kind: 'cartesian', title: 'OHLC', mark: 'candlestick',
-      datasets: [{ id: 'primary', fields: ['label', 'open', 'close', 'low', 'high'].map((id, index) => ({ id, role: index ? 'metric' : 'dimension', dataType: index ? 'decimal' : 'string', nullable: false, label: id })) }],
-      dataBudget: { maxRows: 100, requiredCompleteness: 'complete' }, accessibility: { title: 'OHLC', description: 'OHLC' }, interactions: [],
-      x: { dataset: 'primary', field: 'label' }, y: ['open', 'close', 'low', 'high'].map((field) => ({ dataset: 'primary', field })),
-      presentation: { legend: 'hidden', labelPolicy: { density: 'hidden', priority: [], maxCharacters: 24, minimumSpacing: 0, tooltipFallback: true }, smooth: false, stacked: false, showSymbols: false, dataZoom: true, area: false, step: false },
-    },
-    dataState: { kind: 'inline', specRevision: 'sha256:test', dataRevision: 1, generation: 1, datasets: [
-      { id: 'primary', specRevision: 'sha256:test', dataRevision: 1, generation: 1, columns: ['label', 'open', 'close', 'low', 'high'], rows: [['Jan', 1, 2, 0, 3]], completeness: 'complete' },
-    ] },
-    selection: [], status: { kind: 'ready' }, diagnostics: [],
-  } as VisualizationEnvelope
-
-  const option = echartsOption(envelope) as any
-  expect(option.series).toHaveLength(1)
-  expect(option.xAxis.data).toEqual(['Jan'])
-  expect(option.series[0].encode).toBeUndefined()
-  expect(option.series[0].data).toEqual([{
-    name: 'Jan', value: [1, 2, 0, 3], __lv_dataset: 'primary', __lv_row_index: 0,
-  }])
-})
 
 test('ECharts renders labels inside colored cartesian marks in outlined white', () => {
   const envelope = cartesianFixture('bar') as any
@@ -1419,33 +1143,6 @@ function cartesianFixture(mark: string, columns = ['label', 'value']): Visualiza
     schemaVersion: 9, visualID: mark, rendererID: 'echarts', specRevision: 'sha256:test', dataRevision: 1,
     spec: { kind: 'cartesian', title: mark, mark, datasets: [{ id: 'primary', fields }], dataBudget: { maxRows: 100, requiredCompleteness: 'complete' }, accessibility: { title: mark, description: mark }, interactions: [], x: { dataset: 'primary', field: 'label' }, y, presentation: { legend: 'bottom', labelPolicy: { density: 'automatic', priority: ['selected', 'anomaly', 'threshold'], maxCharacters: 24, minimumSpacing: 6, tooltipFallback: true }, smooth: true, stacked: true, showSymbols: false, dataZoom: true, area: mark === 'area', step: true, symbolSize: 12, labelPosition: 'top', orientation: mark === 'bar' ? 'horizontal' : 'vertical', histogramBins: mark === 'histogram' ? 10 : undefined } },
     dataState: { kind: 'inline', specRevision: 'sha256:test', dataRevision: 1, generation: 1, datasets: [{ id: 'primary', specRevision: 'sha256:test', dataRevision: 1, generation: 1, columns, rows: [row], completeness: 'complete' }] }, selection: [], status: { kind: 'ready' }, diagnostics: [],
-  } as VisualizationEnvelope
-}
-
-function cartesianSeriesFixture(): VisualizationEnvelope {
-  return {
-    schemaVersion: 9, visualID: 'series', rendererID: 'echarts', specRevision: 'sha256:series', dataRevision: 1,
-    spec: {
-      kind: 'cartesian', title: 'Orders', mark: 'area',
-      datasets: [{ id: 'primary', fields: [
-        { id: 'label', role: 'dimension', dataType: 'string', nullable: false, label: 'Month' },
-        { id: 'series', role: 'dimension', dataType: 'string', nullable: false, label: 'Status' },
-        { id: 'value', role: 'metric', dataType: 'decimal', nullable: false, label: 'Orders' },
-      ] }],
-      dataBudget: { maxRows: 100, requiredCompleteness: 'complete' }, accessibility: { title: 'Orders', description: 'Orders by status' }, interactions: [],
-      x: { dataset: 'primary', field: 'label' }, y: [{ dataset: 'primary', field: 'value' }], series: { dataset: 'primary', field: 'series' },
-      presentation: { legend: 'bottom', labelPolicy: { density: 'automatic', priority: ['selected', 'anomaly', 'threshold'], maxCharacters: 24, minimumSpacing: 6, tooltipFallback: true }, smooth: false, stacked: false, stacking: 'normal', showSymbols: true, dataZoom: false, area: true, step: false },
-    },
-    dataState: {
-      kind: 'inline', specRevision: 'sha256:series', dataRevision: 1, generation: 1,
-      datasets: [{
-        id: 'primary', specRevision: 'sha256:series', dataRevision: 1, generation: 1,
-        columns: ['label', 'series', 'value'],
-        rows: [['Jan', 'processing', 30], ['Jan', 'delivered', 10], ['Feb', 'processing', 10], ['Feb', 'delivered', 30]],
-        completeness: 'complete',
-      }],
-    },
-    selection: [], status: { kind: 'ready' }, diagnostics: [],
   } as VisualizationEnvelope
 }
